FW: Update Makefiles to handle sdld errors better and clean output
Now if sdld exit with an error code the make process will detect this.

Also fixed the error where a base address was set for the non-existing
_DATA section.

Switched to a hopefully more portable way to work around the problem,
where echo interprets the '-e' command as an option. The current
solution to use the '-e' was not portable.

Finally hide command output and instead print user friendly strings.